refactor: move 'user' module to 'domain' directory
