+++ src/auth/auth.module.ts
@@ -5,7 +5,7 @@ import { JwtModule } from "@nestjs/jwt";
 import { ConfigService } from "@nestjs/config";
 import { JwtStrategy } from "./strategies/jwt.strategy";
 import { PrismaModule } from "src/prisma/prisma.module";
-import { UserModule } from "src/user/user.module";
+import { UserModule } from "src/domain/user/user.module";
 import { CommonModule } from "src/common/common.module";
 import { RefreshTokenStrategy } from "./strategies/refresh-token.strategy";
