+++ src/domain/user/interfaces/user.ts
