+++ src/domain/user/v1/dto/swagger/deleted-user-api-response.dto.ts
