+++ src/domain/user/v1/dto/swagger/find-user-api.response.ts
